Adds a PATCH route to update event streams
* Event streams could be updated while batches of events are currently
being processed. The existing logic around concurrent handling of batch
processing of events and any pause/stopping of event streams ensures
that batch processing exits as soon as a pause/stop event is detected.
This PR adds an additional context (and a cancel function) to the event
stream spec so that an update to the event stream can signal all the
batch processing related go routines to invoke the cancel function
on the context (via select)
* All fields in the event stream except the ID are allowed to be updated

Addresses one of the items in: https://github.com/kaleido-io/ethconnect/issues/68