ABI: Handle unnamed input parameters in functions (#90)
* ABI: Handle unnamed input parameters in functions

* Solidity supports un-named parameters in functions. The ABI parsing
logic is now enhanced to assign a default name that is prefixed `input`
when processing such functions with unnamed parameters. If there are
multiple unnamed parameters, a suffix with the parameter's index is
added to the name. API requests shall have to use the above nomenclature
when sending requests.
* Request parsing logic is also updated to handle unnamed parameters in
the ABI and look for default names in the query parameters, as described
above.

* Add comments where we are assigning default names

* Remove stale comment

Signed-off-by: Vinod Damle <vinod.damle@kaleido.io>