Adjust highest nonce when there is no gap to fill
When a transaction cannot be sent due to an error (ex: gas estimation
fails), we check and see if a gap exists (inflight txns with higher
nonce exist), in which case a gap fill transaction is sent. However, in
the case where no gap exists, the `highest nonce` for the `from` address
must be reset to one less than the inflight transaction's nonce so that
the same nonce is re-used for the next incoming transaction.